-Adjusted Watergenerator Power Needs. (From 10 to 50 MJ) (Not final yet) -Fixed Powerloss. now its working perfect. I know i knowed the fix already.

+++ Spmod_Common/Source/speiger/src/api/energy/EnergyProvider.java
@@ -143,7 +143,7 @@ public class EnergyProvider
 	{
 		if (storedEnergy > 0)
 		{
-			double loss = storedEnergy / 1000;
+			double loss = (double)storedEnergy / (double)1000;
 			int losedPower = (int) (loss * powerloss);
 			storedEnergy -= losedPower;
 		}
